leaf-disease-backend/requirements.txt
scikit-learn
tensorflow
tensorboard
numpy
pandas
matplotlib
plotly
bs4
beautifulsoup4
selenium
python-dotenv


fastapi
python-multipart
SQLAlchemy
pyodbc
pyswarm
pyswarms

langchain
langchain_community
langchain_core
langchain_google_genai
langchain_openai
langchain_ollama
langchain_huggingface
langchain_chroma

nltk
seaborn
opencv-python